Deep Learning Project/requirements.txt
numpy>=2.0.2
pandas>=2.2.2
scikit-learn>=1.5.1
matplotlib>=3.9.2
seaborn>=0.13.2
nltk>=3.9.1
tensorflow_intel>=2.18.0
keras>=3.7.0
optuna>=3.3.0
kagglehub